refactored code base a bit and updated approach to comparissins to take_run()
