+++ arrow-array/src/lib.rs
@@ -248,6 +248,8 @@ pub use scalar::*;
 
 pub mod builder;
 pub mod cast;
+#[doc(hidden)]
+pub mod comparator;
 mod delta;
 #[cfg(feature = "ffi")]
 pub mod ffi;
